docs(core): make examples compile
Add support in `cdk-build` to actually run `jsii-rosetta --compile --fail`
in packages that declare supporting "strict" mode for `jsii-rosetta`.
This enables a "quick" feedback loop for detercting "now-invalid" code
examples.

Turned on this feature on `@aws-cdk/core` and made the necessary
adjustements to fixtures and examples so that they now compile.